git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/trunk@1272 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CParticleSystemSceneNode.cpp
@@ -422,7 +422,7 @@ void CParticleSystemSceneNode::doParticleSystem(u32 time)
 	// animate all particles
 	f32 scale = (f32)timediff;
 
-	for (s32 i=0; i<(s32)Particles.size();)
+	for (u32 i=0; i<Particles.size();)
 	{
 		if (now > Particles[i].endTime)
 			Particles.erase(i);
@@ -533,7 +533,7 @@ void CParticleSystemSceneNode::serializeAttributes(io::IAttributes* out, io::SAt
 	E_PARTICLE_AFFECTOR_TYPE atype = EPAT_NONE;
 
 	for (core::list<IParticleAffector*>::ConstIterator it = AffectorList.begin();
-		 it != AffectorList.end(); ++it)
+		it != AffectorList.end(); ++it)
 	{
 		atype = (*it)->getType();
